+++ aten/src/ATen/native/vulkan/graph/Functions.cpp
@@ -1,9 +1,10 @@
 #include <ATen/native/vulkan/impl/Arithmetic.h>
 #include <ATen/native/vulkan/impl/Common.h>
 
-#include <ATen/native/vulkan/graph/Arithmetic.h>
 #include <ATen/native/vulkan/graph/Functions.h>
-#include <ATen/native/vulkan/graph/Staging.h>
+
+#include <ATen/native/vulkan/graph/ops/Arithmetic.h>
+#include <ATen/native/vulkan/graph/ops/Staging.h>
 
 namespace at {
 namespace native {
