+++ aten/src/ATen/native/vulkan/graph/Graph.cpp
@@ -1,5 +1,6 @@
 #include <ATen/native/vulkan/graph/Graph.h>
-#include <ATen/native/vulkan/graph/Staging.h>
+
+#include <ATen/native/vulkan/graph/ops/Staging.h>
 
 namespace at {
 namespace native {
